ai-service/requirements.txt
fastapi
python-dotenv
uvicorn
sentence-transformers
playwright
python-jobspy
anthropic
asyncpg
psycopg2-binary
apscheduler
python-multipart
httpx
